Message reading/writing in progress

+++ NetlinkCore/Protocol/SpanReader.cs
@@ -7,21 +7,21 @@ namespace NetlinkCore.Protocol;
 internal unsafe ref struct SpanReader(ReadOnlySpan<byte> buffer)
 {
     private readonly ReadOnlySpan<byte> _buffer = buffer;
-    private int _position = 0;
+    private uint _position = 0;
 
     public readonly bool IsEndOfBuffer => _position == _buffer.Length;
 
     public ReadOnlySpan<byte> Read(int length)
     {
-        var slice = _buffer.Slice(_position, length);
-        _position += Alignment.Align(length);
+        var slice = _buffer.Slice((int)_position, length);
+        _position += Alignment.Align((uint)length);
         return slice;
     }
 
     public ReadOnlySpan<byte> ReadToEnd()
     {
-        var slice = _buffer[_position..];
-        _position = _buffer.Length;
+        var slice = _buffer[(int)_position..];
+        _position = (uint)_buffer.Length;
         return slice;
     }
